Move the logging API
