+++ src/wrappers.rs
@@ -4,6 +4,7 @@ pub mod mj_statistic;
 pub mod mj_rendering;
 pub mod mj_auxiliary;
 pub mod mj_primitive;
+pub mod mj_logging;
 pub mod mj_editing;
 pub mod mj_option;
 pub mod mj_plugin;
@@ -18,6 +19,7 @@ pub use mj_statistic::*;
 pub use mj_rendering::*;
 pub use mj_auxiliary::*;
 pub use mj_primitive::*;
+pub use mj_logging::*;
 pub use mj_plugin::*;
 pub use mj_option::*;
 pub use mj_model::*;
